v0.5: speed up load_ms by numba index; add deep_big cut before fast fdr; remove gc.collect()

+++ beta_dia/fxic.py
@@ -583,7 +583,7 @@ def screen_locus_by_deep(df_batch, locus_num, top_deep_q):
     '''
     group_size_cumsum = np.concatenate([[0], np.cumsum(locus_num)])
     group_rank_deep = utils.cal_group_rank(
-        df_batch['seek_score_deep'].values.astype(np.float32), group_size_cumsum
+        df_batch['seek_score_deep'].values, group_size_cumsum
     )
     group_rank_x = utils.cal_group_rank(
         df_batch['seek_score_sa_x_deep'].values, group_size_cumsum
